+++ package.json
@@ -32,7 +32,6 @@
     "@docusaurus/core": "3.9.2",
     "@docusaurus/preset-classic": "3.9.2",
     "@docusaurus/theme-mermaid": "3.9.2",
-    "@hookform/resolvers": "^5.2.2",
     "@mdx-js/react": "^3.0.0",
     "@modelcontextprotocol/sdk": "^1.25.2",
     "@vercel/analytics": "^2.0.1",
@@ -40,7 +39,6 @@
     "class-variance-authority": "^0.7.1",
     "clsx": "^2.0.0",
     "cmdk": "^1.1.1",
-    "date-fns": "^4.1.0",
     "docusaurus-lunr-search": "3.6.0",
     "embla-carousel-react": "^8.6.0",
     "input-otp": "^1.4.2",
@@ -67,6 +65,7 @@
     "@playwright/test": "^1.58.2",
     "@tailwindcss/postcss": "^4.2.2",
     "@vercel/node": "^5.6.15",
+    "fallow": "^2.45.0",
     "husky": "^9.1.7",
     "image-size": "^2.0.2",
     "postcss": "^8.5.8",
