+++ src/lib/bootstrap-prompt.ts
@@ -5,7 +5,3 @@ export const ABOUT_DEVHUB_SLUG = "about-devhub" as const;
 export function getBootstrapPromptApiPath(): string {
   return `/api/bootstrap-prompt`;
 }
-
-export function getRecipeMarkdownApiPath(): string {
-  return `/api/markdown?section=${BOOTSTRAP_PROMPT_SECTION}&slug=${BOOTSTRAP_PROMPT_SLUG}`;
-}
