+++ src/pages/templates/ai-chat-app.tsx
@@ -1,11 +1,6 @@
 import type { ReactNode } from "react";
 import { CookbookDetail } from "@/components/cookbooks/cookbook-detail";
-import { cookbooks, recipes } from "@/lib/recipes/recipes";
-import {
-  useAllRecipeSections,
-  useCookbookIntro,
-} from "@/lib/use-raw-content-markdown";
-import { composeCookbookMarkdown } from "@/lib/cookbook-composition";
+import { useCookbookMarkdown } from "@/lib/use-cookbook-markdown";
 import Intro from "@site/content/cookbooks/ai-chat-app/intro.md";
 import BootstrapPrereqs from "@site/content/recipes/databricks-local-bootstrap/prerequisites.md";
 import BootstrapContent from "@site/content/recipes/databricks-local-bootstrap/content.md";
@@ -20,30 +15,8 @@ import LakebaseDataPersistenceContent from "@site/content/recipes/lakebase-data-
 import LakebaseChatPersistencePrereqs from "@site/content/recipes/lakebase-chat-persistence/prerequisites.md";
 import LakebaseChatPersistenceContent from "@site/content/recipes/lakebase-chat-persistence/content.md";
 
-const COOKBOOK_ID = "ai-chat-app";
-
 export default function AiChatAppPage(): ReactNode {
-  const cookbook = cookbooks.find((t) => t.id === COOKBOOK_ID);
-  if (!cookbook) throw new Error(`Cookbook ${COOKBOOK_ID} not found`);
-
-  const sectionsBySlug = useAllRecipeSections();
-  const intro = useCookbookIntro(COOKBOOK_ID);
-
-  const recipeInputs = cookbook.recipeIds.map((id) => {
-    const recipe = recipes.find((r) => r.id === id);
-    const sections = sectionsBySlug[id];
-    if (!recipe || !sections) {
-      throw new Error(`Missing recipe or sections for "${id}"`);
-    }
-    return { id, name: recipe.name, sections };
-  });
-
-  const rawMarkdown = composeCookbookMarkdown({
-    cookbookName: cookbook.name,
-    cookbookDescription: cookbook.description,
-    intro,
-    recipes: recipeInputs,
-  });
+  const { cookbook, rawMarkdown } = useCookbookMarkdown("ai-chat-app");
 
   return (
     <CookbookDetail cookbook={cookbook} rawMarkdown={rawMarkdown}>
