+++ src/pages/templates/hello-world-app.tsx
@@ -1,38 +1,11 @@
 import type { ReactNode } from "react";
 import { CookbookDetail } from "@/components/cookbooks/cookbook-detail";
-import { cookbooks, recipes } from "@/lib/recipes/recipes";
-import {
-  useAllRecipeSections,
-  useCookbookIntro,
-} from "@/lib/use-raw-content-markdown";
-import { composeCookbookMarkdown } from "@/lib/cookbook-composition";
+import { useCookbookMarkdown } from "@/lib/use-cookbook-markdown";
 import BootstrapPrereqs from "@site/content/recipes/databricks-local-bootstrap/prerequisites.md";
 import BootstrapContent from "@site/content/recipes/databricks-local-bootstrap/content.md";
 
-const COOKBOOK_ID = "hello-world-app";
-
 export default function HelloWorldAppPage(): ReactNode {
-  const cookbook = cookbooks.find((t) => t.id === COOKBOOK_ID);
-  if (!cookbook) throw new Error(`Cookbook ${COOKBOOK_ID} not found`);
-
-  const sectionsBySlug = useAllRecipeSections();
-  const intro = useCookbookIntro(COOKBOOK_ID);
-
-  const recipeInputs = cookbook.recipeIds.map((id) => {
-    const recipe = recipes.find((r) => r.id === id);
-    const sections = sectionsBySlug[id];
-    if (!recipe || !sections) {
-      throw new Error(`Missing recipe or sections for "${id}"`);
-    }
-    return { id, name: recipe.name, sections };
-  });
-
-  const rawMarkdown = composeCookbookMarkdown({
-    cookbookName: cookbook.name,
-    cookbookDescription: cookbook.description,
-    intro,
-    recipes: recipeInputs,
-  });
+  const { cookbook, rawMarkdown } = useCookbookMarkdown("hello-world-app");
 
   return (
     <CookbookDetail cookbook={cookbook} rawMarkdown={rawMarkdown}>
