+++ src/pages/templates/lakebase-off-platform.tsx
@@ -1,11 +1,6 @@
 import type { ReactNode } from "react";
 import { CookbookDetail } from "@/components/cookbooks/cookbook-detail";
-import { cookbooks, recipes } from "@/lib/recipes/recipes";
-import {
-  useAllRecipeSections,
-  useCookbookIntro,
-} from "@/lib/use-raw-content-markdown";
-import { composeCookbookMarkdown } from "@/lib/cookbook-composition";
+import { useCookbookMarkdown } from "@/lib/use-cookbook-markdown";
 import LakebaseCreateInstancePrereqs from "@site/content/recipes/lakebase-create-instance/prerequisites.md";
 import LakebaseCreateInstanceContent from "@site/content/recipes/lakebase-create-instance/content.md";
 import LakebaseOffPlatformEnvManagementPrereqs from "@site/content/recipes/lakebase-off-platform-env-management/prerequisites.md";
@@ -15,30 +10,10 @@ import LakebaseTokenManagementContent from "@site/content/recipes/lakebase-token
 import LakebaseDrizzleOffPlatformPrereqs from "@site/content/recipes/lakebase-drizzle-off-platform/prerequisites.md";
 import LakebaseDrizzleOffPlatformContent from "@site/content/recipes/lakebase-drizzle-off-platform/content.md";
 
-const COOKBOOK_ID = "lakebase-off-platform";
-
 export default function LakebaseOffPlatformPage(): ReactNode {
-  const cookbook = cookbooks.find((t) => t.id === COOKBOOK_ID);
-  if (!cookbook) throw new Error(`Cookbook ${COOKBOOK_ID} not found`);
-
-  const sectionsBySlug = useAllRecipeSections();
-  const intro = useCookbookIntro(COOKBOOK_ID);
-
-  const recipeInputs = cookbook.recipeIds.map((id) => {
-    const recipe = recipes.find((r) => r.id === id);
-    const sections = sectionsBySlug[id];
-    if (!recipe || !sections) {
-      throw new Error(`Missing recipe or sections for "${id}"`);
-    }
-    return { id, name: recipe.name, sections };
-  });
-
-  const rawMarkdown = composeCookbookMarkdown({
-    cookbookName: cookbook.name,
-    cookbookDescription: cookbook.description,
-    intro,
-    recipes: recipeInputs,
-  });
+  const { cookbook, rawMarkdown } = useCookbookMarkdown(
+    "lakebase-off-platform",
+  );
 
   return (
     <CookbookDetail cookbook={cookbook} rawMarkdown={rawMarkdown}>
